Update #3
- Some BugFix
- Refactoring

+++ User.cpp
@@ -2,13 +2,14 @@
 
 
 
-User::User(string Name, unsigned long long int Money)
+User::User(string Name, uint64_t Money)
 {
 	this->Name = Name;
 	this->Money = Money;
 }
 
-unsigned long long int User::GetMoney() {
+
+uint64_t User::GetMoney() {
 	return this->Money;
 }
 
@@ -16,6 +17,6 @@ string User::GetName() {
 	return this->Name;
 }
 
-void User::SetMoney(unsigned long long int Money) {
+void User::SetMoney(uint64_t Money) {
 	this->Money = Money;
 }
